// apps/cli/src/types/supabase.ts
export type Json =
  | string
  | number
  | boolean
  | null
  | { [key: string]: Json | undefined }
  | Json[]

export type Database = {
// ...
  public: {
    Tables: {
      epics: {
        Row: {
          color_code: string
          created_at: string
          id: string
          name: string
          user_id: string
        }
        Insert: {
          color_code: string
          created_at?: string
          id?: string
          name: string
          user_id: string
        }
        Update: {
          color_code?: string
          created_at?: string
          id?: string
          name?: string
          user_id?: string
        }
        Relationships: []
      }
      projects: {
        Row: {
          color_code: string | null
          created_at: string
          epic_id: string
          id: string
          name: string
          status: Database["public"]["Enums"]["project_status"]
          user_id: string
        }
        Insert: {
          color_code?: string | null
          created_at?: string
          epic_id: string
          id?: string
          name: string
          status?: Database["public"]["Enums"]["project_status"]
          user_id: string
        }
        Update: {
          color_code?: string | null
          created_at?: string
          epic_id?: string
          id?: string
          name?: string
          status?: Database["public"]["Enums"]["project_status"]
          user_id?: string
        }
        Relationships: [
          {
            foreignKeyName: "projects_epic_id_fkey"
            columns: ["epic_id"]
            isOneToOne: false
            referencedRelation: "epics"
            referencedColumns: ["id"]
          },
        ]
      }
// ...
      todos: {
        Row: {
          completed_at: string | null
          content: Json
          content_format: string
          created_at: string
          expiration_date: string | null
          id: string
          priority: Database["public"]["Enums"]["task_priority"]
          project_id: string
          status: Database["public"]["Enums"]["todo_status"]
          title: string
          updated_at: string
          urgency: Database["public"]["Enums"]["task_priority"]
          user_id: string
        }
        Insert: {
          completed_at?: string | null
          content?: Json
          content_format?: string
          created_at?: string
          expiration_date?: string | null
          id?: string
          priority?: Database["public"]["Enums"]["task_priority"]
          project_id: string
          status?: Database["public"]["Enums"]["todo_status"]
          title: string
          updated_at?: string
          urgency?: Database["public"]["Enums"]["task_priority"]
          user_id: string
        }
        Update: {
          completed_at?: string | null
          content?: Json
          content_format?: string
          created_at?: string
          expiration_date?: string | null
          id?: string
          priority?: Database["public"]["Enums"]["task_priority"]
          project_id?: string
          status?: Database["public"]["Enums"]["todo_status"]
          title?: string
          updated_at?: string
          urgency?: Database["public"]["Enums"]["task_priority"]
          user_id?: string
        }
        Relationships: [
          {
            foreignKeyName: "todos_project_id_fkey"
            columns: ["project_id"]
            isOneToOne: false
            referencedRelation: "projects"
            referencedColumns: ["id"]
          },
        ]
      }
    }
// ...
  }
}

type DatabaseWithoutInternals = Omit<Database, "__InternalSupabase">

type DefaultSchema = DatabaseWithoutInternals[Extract<keyof Database, "public">]
// ...
export type Enums<
  DefaultSchemaEnumNameOrOptions extends
    | keyof DefaultSchema["Enums"]
    | { schema: keyof DatabaseWithoutInternals },
  EnumName extends (DefaultSchemaEnumNameOrOptions extends {
    schema: keyof DatabaseWithoutInternals
  }
    ? keyof DatabaseWithoutInternals[DefaultSchemaEnumNameOrOptions["schema"]]["Enums"]
    : never) = never,
> = DefaultSchemaEnumNameOrOptions extends {
  schema: keyof DatabaseWithoutInternals
}
  ? DatabaseWithoutInternals[DefaultSchemaEnumNameOrOptions["schema"]]["Enums"][EnumName]
  : DefaultSchemaEnumNameOrOptions extends keyof DefaultSchema["Enums"]
    ? DefaultSchema["Enums"][DefaultSchemaEnumNameOrOptions]
    : never